+++ web/default/src/components/data-table/data-table-page.tsx
@@ -342,7 +342,10 @@ function renderDesktop<TData>(
                 <TableHead
                   key={header.id}
                   colSpan={header.colSpan}
-                  className={header.column.columnDef.meta?.className}
+                  className={cn(
+                    header.column.columnDef.meta?.className,
+                    header.column.columnDef.meta?.headerClassName
+                  )}
                   style={
                     props.applyHeaderSize
                       ? { width: header.getSize() }
@@ -410,7 +413,10 @@ function DefaultRow<TData>({
       {row.getVisibleCells().map((cell) => (
         <TableCell
           key={cell.id}
-          className={cell.column.columnDef.meta?.className}
+          className={cn(
+            cell.column.columnDef.meta?.className,
+            cell.column.columnDef.meta?.cellClassName
+          )}
         >
           {flexRender(cell.column.columnDef.cell, cell.getContext())}
         </TableCell>
