+++ web/default/src/features/channels/lib/channel-config-validation.ts
@@ -17,14 +17,14 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 For commercial licensing, please contact https://github.com/MAX-API-Next/MAX-API/issues
 */
 import { MODEL_FETCHABLE_TYPES } from '../constants'
-import type { ChannelFormValues } from './channel-form'
+import { isVideoTaskChannelType } from './channel-capabilities'
 import {
   BASE_URL_REQUIRED_TYPES,
   OTHER_REQUIRED_TYPES,
   hasVertexDefaultRegion,
   hasVideoTaskQueryPlaceholder,
 } from './channel-config-rules'
-import { isVideoTaskChannelType } from './channel-capabilities'
+import type { ChannelFormValues } from './channel-form'
 
 export type ChannelConfigValidationSeverity = 'error' | 'warning' | 'info'
 
@@ -87,7 +87,7 @@ function validateJsonFields(
         id: `${field}_invalid_json`,
         severity: 'error',
         field,
-        message: '此 JSON 配置必须是 JSON 对象。',
+        message: 'This JSON configuration must be a JSON object.',
       })
     }
   }
@@ -107,12 +107,13 @@ function validateVertexConfig(
   try {
     regionConfig = parseJson(otherValue)
   } catch {
-      addIssue(issues, {
-        id: 'vertex_region_invalid_json',
-        severity: 'error',
-        field: 'other',
-        message: 'Vertex AI 区域配置必须是包含 default 字段的 JSON 对象。',
-      })
+    addIssue(issues, {
+      id: 'vertex_region_invalid_json',
+      severity: 'error',
+      field: 'other',
+      message:
+        'Vertex AI region configuration must be a JSON object containing a default field.',
+    })
     return
   }
 
@@ -121,7 +122,7 @@ function validateVertexConfig(
       id: 'vertex_region_missing_default',
       severity: 'error',
       field: 'other',
-      message: 'Vertex AI 区域配置必须包含 default 字段。',
+      message: 'Vertex AI region configuration must contain a default field.',
     })
   }
 
@@ -131,7 +132,7 @@ function validateVertexConfig(
         id: 'vertex_api_key_batch_mode',
         severity: 'error',
         field: 'multi_key_mode',
-        message: 'Vertex AI 的 API Key 模式不支持批量创建。',
+        message: 'Vertex AI API Key mode does not support batch creation.',
       })
     }
     return
@@ -148,7 +149,7 @@ function validateVertexConfig(
         id: 'vertex_key_invalid_shape',
         severity: 'error',
         field: 'key',
-        message: 'Vertex AI 服务账号密钥必须是有效的 JSON。',
+        message: 'Vertex AI service account credentials must be valid JSON.',
       })
     }
   } catch {
@@ -156,7 +157,7 @@ function validateVertexConfig(
       id: 'vertex_key_invalid_json',
       severity: 'error',
       field: 'key',
-      message: 'Vertex AI 服务账号密钥必须是有效的 JSON。',
+      message: 'Vertex AI service account credentials must be valid JSON.',
     })
   }
 }
@@ -173,7 +174,7 @@ function validateCodexConfig(
       id: 'codex_batch_mode',
       severity: 'error',
       field: 'multi_key_mode',
-      message: 'Codex 渠道不支持批量创建。',
+      message: 'Codex channels do not support batch creation.',
     })
   }
 
@@ -192,7 +193,7 @@ function validateCodexConfig(
         id: 'codex_key_missing_fields',
         severity: 'error',
         field: 'key',
-        message: 'Codex 凭证必须包含 access_token 和 account_id 字段。',
+        message: 'Codex credentials must contain access_token and account_id.',
       })
     }
   } catch {
@@ -200,7 +201,7 @@ function validateCodexConfig(
       id: 'codex_key_invalid_json',
       severity: 'error',
       field: 'key',
-      message: 'Codex 凭证必须是有效的 JSON 对象。',
+      message: 'Codex credentials must be a valid JSON object.',
     })
   }
 }
@@ -218,7 +219,8 @@ function validateVideoTaskConfig(
       id: 'video_task_unsupported_type',
       severity: 'warning',
       field: 'video_task_protocol_enabled',
-      message: '视频任务协议设置仅适用于支持视频任务的渠道类型。',
+      message:
+        'Video task protocol settings are only available for channel types that support video tasks.',
     })
     return
   }
@@ -235,7 +237,7 @@ function validateVideoTaskConfig(
       id: 'video_task_missing_submit_path',
       severity: 'error',
       field: 'video_task_submit_path',
-      message: '视频任务提交路径不能为空。',
+      message: 'Video task submission path is required.',
     })
   }
 
@@ -245,7 +247,7 @@ function validateVideoTaskConfig(
       id: 'video_task_missing_query_path',
       severity: 'error',
       field: 'video_task_query_path',
-      message: '视频任务查询路径不能为空。',
+      message: 'Video task query path is required.',
     })
   } else if (!hasVideoTaskQueryPlaceholder(queryPath)) {
     addIssue(issues, {
@@ -253,7 +255,7 @@ function validateVideoTaskConfig(
       severity: 'error',
       field: 'video_task_query_path',
       message:
-        '视频任务查询路径必须包含 {task_id}、{operation_name} 或 {upstream_task_id}。',
+        'Video task query path must contain {task_id}, {operation_name}, or {upstream_task_id}.',
     })
   }
 
@@ -263,7 +265,8 @@ function validateVideoTaskConfig(
         id: 'video_task_missing_task_id_path',
         severity: 'error',
         field: 'video_task_task_id_path',
-        message: '启用完整任务协议时，必须填写任务 ID 响应路径。',
+        message:
+          'Task ID response path is required when the full task protocol is enabled.',
       })
     }
     if (!trim(values.video_task_status_path)) {
@@ -271,7 +274,8 @@ function validateVideoTaskConfig(
         id: 'video_task_missing_status_path',
         severity: 'error',
         field: 'video_task_status_path',
-        message: '启用完整任务协议时，必须填写状态响应路径。',
+        message:
+          'Status response path is required when the full task protocol is enabled.',
       })
     }
     if (!trim(values.video_task_result_url_paths)) {
@@ -279,7 +283,8 @@ function validateVideoTaskConfig(
         id: 'video_task_missing_result_paths',
         severity: 'warning',
         field: 'video_task_result_url_paths',
-        message: '未配置结果 URL 路径，视频任务完成后可能无法返回内容地址。',
+        message:
+          'Without a result URL path, completed video tasks may not return a content URL.',
       })
     }
   }
@@ -297,7 +302,7 @@ export function getChannelConfigValidationIssues(
       id: 'key_required_create',
       severity: 'error',
       field: 'key',
-      message: '新建渠道时 API Key 不能为空。',
+      message: 'API Key is required when creating a channel.',
     })
   }
 
@@ -306,7 +311,7 @@ export function getChannelConfigValidationIssues(
       id: 'models_required',
       severity: 'error',
       field: 'models',
-      message: '该渠道至少需要发布一个模型。',
+      message: 'At least one model must be published for this channel.',
     })
   }
 
@@ -315,7 +320,7 @@ export function getChannelConfigValidationIssues(
       id: 'base_url_required',
       severity: 'error',
       field: 'base_url',
-      message: '当前渠道类型需要填写 Base URL。',
+      message: 'Base URL is required for this channel type.',
     })
   }
 
@@ -325,7 +330,7 @@ export function getChannelConfigValidationIssues(
       severity: 'warning',
       field: 'base_url',
       message:
-        'Base URL 通常不应以 /v1 结尾，MAX API 会自动拼接上游路径。',
+        'Base URL should not usually end with /v1 because MAX API appends the upstream path automatically.',
     })
   }
 
@@ -334,7 +339,8 @@ export function getChannelConfigValidationIssues(
       id: 'other_required',
       severity: 'error',
       field: 'other',
-      message: '当前渠道类型需要填写额外的上游配置。',
+      message:
+        'Additional upstream configuration is required for this channel type.',
     })
   }
 
@@ -347,7 +353,8 @@ export function getChannelConfigValidationIssues(
       id: 'model_discovery_not_supported',
       severity: 'warning',
       field: 'upstream_model_update_check_enabled',
-      message: '当前渠道类型不支持上游模型发现，该功能不会运行。',
+      message:
+        'Upstream model discovery is not supported for this channel type and will not run.',
     })
   }
