+++ web/default/src/features/system-settings/models/tiered-pricing-request-rules.tsx
@@ -1,0 +1,510 @@
+﻿/*
+Copyright (C) 2023-2026 MAX-API-Next
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU Affero General Public License as
+published by the Free Software Foundation, either version 3 of the
+License, or (at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+GNU Affero General Public License for more details.
+
+You should have received a copy of the GNU Affero General Public License
+along with this program. If not, see <https://www.gnu.org/licenses/>.
+
+For commercial licensing, please contact https://github.com/MAX-API-Next/MAX-API/issues
+*/
+import { useMemo } from 'react'
+import { Plus, Trash2 } from 'lucide-react'
+import { useTranslation } from 'react-i18next'
+import { Alert, AlertDescription } from '@/components/ui/alert'
+import { Badge } from '@/components/ui/badge'
+import { Button } from '@/components/ui/button'
+import { Input } from '@/components/ui/input'
+import { Label } from '@/components/ui/label'
+import {
+  Select,
+  SelectContent,
+  SelectGroup,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from '@/components/ui/select'
+import {
+  COMMON_TIMEZONES,
+  MATCH_CONTAINS,
+  MATCH_EQ,
+  MATCH_EXISTS,
+  MATCH_GT,
+  MATCH_GTE,
+  MATCH_LT,
+  MATCH_LTE,
+  MATCH_RANGE,
+  SOURCE_HEADER,
+  SOURCE_PARAM,
+  SOURCE_TIME,
+  TIME_FUNCS,
+  createEmptyCondition,
+  createEmptyRuleGroup,
+  createEmptyTimeCondition,
+  getRequestRuleMatchOptions,
+  tryParseRequestRuleExpr,
+  type ParamHeaderCondition,
+  type RequestCondition,
+  type RequestRuleGroup,
+  type TimeCondition,
+  type TimeFunc,
+} from '@/features/pricing/lib/billing-expr'
+import { DraftNumberInput } from './tiered-pricing-fields'
+
+// ---------------------------------------------------------------------------
+// Request rule condition row
+// ---------------------------------------------------------------------------
+
+type RuleConditionRowProps = {
+  condition: RequestCondition
+  onChange: (next: RequestCondition) => void
+  onRemove: () => void
+}
+
+function RuleConditionRow({
+  condition,
+  onChange,
+  onRemove,
+}: RuleConditionRowProps) {
+  const { t } = useTranslation()
+  const matchOptions = getRequestRuleMatchOptions(condition.source)
+  const getMatchLabel = (mode: string) => {
+    switch (mode) {
+      case MATCH_EQ:
+        return t('Equals')
+      case MATCH_CONTAINS:
+        return t('Contains')
+      case MATCH_EXISTS:
+        return t('Exists')
+      case MATCH_GT:
+        return t('Greater than')
+      case MATCH_GTE:
+        return t('Greater than or equal')
+      case MATCH_LT:
+        return t('Less than')
+      case MATCH_LTE:
+        return t('Less than or equal')
+      case MATCH_RANGE:
+        return t('Overnight range')
+      default:
+        return mode
+    }
+  }
+  const getTimeFuncLabel = (timeFunc: TimeFunc) => {
+    switch (timeFunc) {
+      case 'hour':
+        return t('Hour of day')
+      case 'minute':
+        return t('Minute')
+      case 'weekday':
+        return t('Weekday')
+      case 'month':
+        return t('Month number')
+      case 'day':
+        return t('Day of month')
+      default:
+        return timeFunc
+    }
+  }
+  const sourceLabel =
+    condition.source === SOURCE_PARAM
+      ? t('Body param')
+      : condition.source === SOURCE_HEADER
+        ? t('Header')
+        : t('Time')
+
+  const handleSourceChange = (source: string) => {
+    if (source === SOURCE_TIME) {
+      onChange(createEmptyTimeCondition())
+    } else if (source === SOURCE_HEADER || source === SOURCE_PARAM) {
+      onChange({
+        ...createEmptyCondition(),
+        source: source as 'param' | 'header',
+      })
+    }
+  }
+
+  const handleModeChange = (mode: string) => {
+    onChange({ ...condition, mode } as RequestCondition)
+  }
+
+  const renderTimeCondition = (timeCond: TimeCondition) => (
+    <>
+      <Select
+        items={[
+          ...TIME_FUNCS.map((fn) => ({
+            value: fn,
+            label: getTimeFuncLabel(fn),
+          })),
+        ]}
+        value={timeCond.timeFunc}
+        onValueChange={(value) =>
+          onChange({ ...timeCond, timeFunc: value as TimeFunc })
+        }
+      >
+        <SelectTrigger className='w-32' size='sm'>
+          <SelectValue>{getTimeFuncLabel(timeCond.timeFunc)}</SelectValue>
+        </SelectTrigger>
+        <SelectContent alignItemWithTrigger={false}>
+          <SelectGroup>
+            {TIME_FUNCS.map((fn) => (
+              <SelectItem key={fn} value={fn}>
+                {getTimeFuncLabel(fn)}
+              </SelectItem>
+            ))}
+          </SelectGroup>
+        </SelectContent>
+      </Select>
+      <Select
+        items={[
+          ...COMMON_TIMEZONES.map((tz) => ({
+            value: tz.value,
+            label: tz.label,
+          })),
+        ]}
+        value={timeCond.timezone}
+        onValueChange={(value) =>
+          value !== null && onChange({ ...timeCond, timezone: value })
+        }
+      >
+        <SelectTrigger className='w-56' size='sm'>
+          <SelectValue>
+            {COMMON_TIMEZONES.find((tz) => tz.value === timeCond.timezone)
+              ?.label ?? timeCond.timezone}
+          </SelectValue>
+        </SelectTrigger>
+        <SelectContent alignItemWithTrigger={false}>
+          <SelectGroup>
+            {COMMON_TIMEZONES.map((tz) => (
+              <SelectItem key={tz.value} value={tz.value}>
+                {tz.label}
+              </SelectItem>
+            ))}
+          </SelectGroup>
+        </SelectContent>
+      </Select>
+      <Select
+        items={[
+          ...matchOptions.map((option) => ({
+            value: option.value,
+            label: getMatchLabel(option.value),
+          })),
+        ]}
+        value={timeCond.mode}
+        onValueChange={(v) => v !== null && handleModeChange(v)}
+      >
+        <SelectTrigger className='w-32' size='sm'>
+          <SelectValue>{getMatchLabel(timeCond.mode)}</SelectValue>
+        </SelectTrigger>
+        <SelectContent alignItemWithTrigger={false}>
+          <SelectGroup>
+            {matchOptions.map((option) => (
+              <SelectItem key={option.value} value={option.value}>
+                {getMatchLabel(option.value)}
+              </SelectItem>
+            ))}
+          </SelectGroup>
+        </SelectContent>
+      </Select>
+      {timeCond.mode === MATCH_RANGE ? (
+        <>
+          <DraftNumberInput
+            value={timeCond.rangeStart}
+            onValueChange={(value) =>
+              onChange({ ...timeCond, rangeStart: String(value) })
+            }
+            placeholder={t('Start')}
+            className='w-20'
+          />
+          <span className='text-muted-foreground text-xs'>~</span>
+          <DraftNumberInput
+            value={timeCond.rangeEnd}
+            onValueChange={(value) =>
+              onChange({ ...timeCond, rangeEnd: String(value) })
+            }
+            placeholder={t('End')}
+            className='w-20'
+          />
+        </>
+      ) : (
+        <DraftNumberInput
+          value={timeCond.value}
+          onValueChange={(value) =>
+            onChange({ ...timeCond, value: String(value) })
+          }
+          placeholder={t('Value')}
+          className='w-24'
+        />
+      )}
+    </>
+  )
+
+  const renderParamHeaderCondition = (phCond: ParamHeaderCondition) => (
+    <>
+      <Input
+        value={phCond.path}
+        onChange={(event) => onChange({ ...phCond, path: event.target.value })}
+        placeholder={
+          phCond.source === SOURCE_HEADER ? 'X-Header-Name' : 'service_tier'
+        }
+        className='w-44'
+      />
+      <Select
+        items={[
+          ...matchOptions.map((option) => ({
+            value: option.value,
+            label: getMatchLabel(option.value),
+          })),
+        ]}
+        value={phCond.mode}
+        onValueChange={(v) => v !== null && handleModeChange(v)}
+      >
+        <SelectTrigger className='w-32' size='sm'>
+          <SelectValue>{getMatchLabel(phCond.mode)}</SelectValue>
+        </SelectTrigger>
+        <SelectContent alignItemWithTrigger={false}>
+          <SelectGroup>
+            {matchOptions.map((option) => (
+              <SelectItem key={option.value} value={option.value}>
+                {getMatchLabel(option.value)}
+              </SelectItem>
+            ))}
+          </SelectGroup>
+        </SelectContent>
+      </Select>
+      {phCond.mode !== MATCH_EXISTS && (
+        <Input
+          value={phCond.value}
+          onChange={(event) =>
+            onChange({ ...phCond, value: event.target.value })
+          }
+          placeholder={t('Value')}
+          className='w-44'
+        />
+      )}
+    </>
+  )
+
+  return (
+    <div className='flex flex-wrap items-center gap-2'>
+      <Select
+        items={[
+          { value: SOURCE_PARAM, label: t('Body param') },
+          { value: SOURCE_HEADER, label: t('Header') },
+          { value: SOURCE_TIME, label: t('Time') },
+        ]}
+        value={condition.source}
+        onValueChange={(v) => v !== null && handleSourceChange(v)}
+      >
+        <SelectTrigger className='w-28' size='sm'>
+          <SelectValue>{sourceLabel}</SelectValue>
+        </SelectTrigger>
+        <SelectContent alignItemWithTrigger={false}>
+          <SelectGroup>
+            <SelectItem value={SOURCE_PARAM}>{t('Body param')}</SelectItem>
+            <SelectItem value={SOURCE_HEADER}>{t('Header')}</SelectItem>
+            <SelectItem value={SOURCE_TIME}>{t('Time')}</SelectItem>
+          </SelectGroup>
+        </SelectContent>
+      </Select>
+      {condition.source === SOURCE_TIME
+        ? renderTimeCondition(condition as TimeCondition)
+        : renderParamHeaderCondition(condition as ParamHeaderCondition)}
+      <Button
+        variant='ghost'
+        size='icon'
+        onClick={onRemove}
+        aria-label={t('Remove condition')}
+        className='ml-auto'
+      >
+        <Trash2 className='text-destructive h-4 w-4' />
+      </Button>
+    </div>
+  )
+}
+
+// ---------------------------------------------------------------------------
+// Request rule group card
+// ---------------------------------------------------------------------------
+
+type RuleGroupCardProps = {
+  group: RequestRuleGroup
+  index: number
+  onChange: (next: RequestRuleGroup) => void
+  onRemove: () => void
+}
+
+function RuleGroupCard({
+  group,
+  index,
+  onChange,
+  onRemove,
+}: RuleGroupCardProps) {
+  const { t } = useTranslation()
+
+  const handleConditionChange = (
+    conditionIndex: number,
+    next: RequestCondition
+  ) => {
+    const conditions = [...group.conditions]
+    conditions[conditionIndex] = next
+    onChange({ ...group, conditions })
+  }
+
+  const handleAddCondition = (timeMode: boolean) => {
+    onChange({
+      ...group,
+      conditions: [
+        ...group.conditions,
+        timeMode ? createEmptyTimeCondition() : createEmptyCondition(),
+      ],
+    })
+  }
+
+  return (
+    <div className='bg-muted/30 space-y-3 rounded-md border p-3'>
+      <div className='flex items-center justify-between gap-2'>
+        <Badge variant='outline'>
+          {t('Rule group')} #{index + 1}
+        </Badge>
+        <Button
+          variant='ghost'
+          size='icon'
+          onClick={onRemove}
+          aria-label={t('Remove rule group')}
+        >
+          <Trash2 className='text-destructive h-4 w-4' />
+        </Button>
+      </div>
+
+      <div className='space-y-2'>
+        {group.conditions.map((condition, conditionIndex) => (
+          <RuleConditionRow
+            key={conditionIndex}
+            condition={condition}
+            onChange={(next) => handleConditionChange(conditionIndex, next)}
+            onRemove={() =>
+              onChange({
+                ...group,
+                conditions: group.conditions.filter(
+                  (_, i) => i !== conditionIndex
+                ),
+              })
+            }
+          />
+        ))}
+        <div className='flex flex-wrap gap-2'>
+          <Button
+            variant='ghost'
+            size='sm'
+            onClick={() => handleAddCondition(false)}
+          >
+            <Plus className='mr-1 h-3 w-3' />
+            {t('Add param/header')}
+          </Button>
+          <Button
+            variant='ghost'
+            size='sm'
+            onClick={() => handleAddCondition(true)}
+          >
+            <Plus className='mr-1 h-3 w-3' />
+            {t('Add time condition')}
+          </Button>
+        </div>
+      </div>
+
+      <div className='flex items-center gap-2'>
+        <Label className='text-xs'>{t('Multiplier')}</Label>
+        <DraftNumberInput
+          min={0}
+          step={0.000001}
+          value={group.multiplier}
+          onValueChange={(value) =>
+            onChange({ ...group, multiplier: String(value) })
+          }
+          className='w-32'
+          placeholder='1.0'
+        />
+        <span className='text-muted-foreground text-xs'>
+          {t('Final cost = base × multiplier when conditions match')}
+        </span>
+      </div>
+    </div>
+  )
+}
+
+type RequestRuleEditorProps = {
+  requestRuleExpr: string
+  groups: RequestRuleGroup[]
+  onChange: (next: RequestRuleGroup[]) => void
+}
+
+export function RequestRuleEditor({
+  requestRuleExpr,
+  groups,
+  onChange,
+}: RequestRuleEditorProps) {
+  const { t } = useTranslation()
+  const canUseVisualRules = useMemo(() => {
+    if (!requestRuleExpr) return true
+    return tryParseRequestRuleExpr(requestRuleExpr) !== null
+  }, [requestRuleExpr])
+
+  return (
+    <div className='space-y-3 border-t pt-3'>
+      <div className='space-y-1'>
+        <h4 className='text-sm font-medium'>{t('Request rule pricing')}</h4>
+        <p className='text-muted-foreground text-xs'>
+          {t(
+            'When conditions match, the final price is multiplied by X. Multiple matches multiply together; values < 1 act as discounts.'
+          )}
+        </p>
+      </div>
+
+      {requestRuleExpr && !canUseVisualRules ? (
+        <Alert>
+          <AlertDescription className='text-xs'>
+            {t(
+              'This expression is too complex for the visual editor. Please switch to expression mode to edit.'
+            )}
+          </AlertDescription>
+        </Alert>
+      ) : (
+        <>
+          {groups.map((group, groupIndex) => (
+            <RuleGroupCard
+              key={groupIndex}
+              group={group}
+              index={groupIndex}
+              onChange={(next) => {
+                const updated = [...groups]
+                updated[groupIndex] = next
+                onChange(updated)
+              }}
+              onRemove={() =>
+                onChange(groups.filter((_, index) => index !== groupIndex))
+              }
+            />
+          ))}
+          <Button
+            variant='outline'
+            size='sm'
+            className='h-9 w-36 justify-center'
+            onClick={() => onChange([...groups, createEmptyRuleGroup()])}
+          >
+            <Plus className='mr-2 h-4 w-4' />
+            {t('Add rule group')}
+          </Button>
+        </>
+      )}
+    </div>
+  )
+}
