+++ web/default/src/i18n/locales/en.json
@@ -3,6 +3,58 @@
     "360": "360",
     "1000": "1000",
     "10000": "10000",
+    "{{supported}}/{{total}} available": "{{supported}}/{{total}} available",
+    "Shows the request APIs, task capabilities, and operational features supported by this channel type.": "Shows the request APIs, task capabilities, and operational features supported by this channel type.",
+    "{{count}} blocking issue(s)": "{{count}} blocking issue(s)",
+    "Configuration valid": "Configuration valid",
+    "{{count}} warning(s)": "{{count}} warning(s)",
+    "Configuration validation": "Configuration validation",
+    "No blocking configuration issues were detected for this channel type.": "No blocking configuration issues were detected for this channel type.",
+    "Routes OpenAI-compatible chat and completion requests.": "Routes OpenAI-compatible chat and completion requests.",
+    "Routes OpenAI Responses API requests.": "Routes OpenAI Responses API requests.",
+    "Routes Anthropic Messages-compatible requests.": "Routes Anthropic Messages-compatible requests.",
+    "Routes Gemini native /v1beta model requests.": "Routes Gemini native /v1beta model requests.",
+    "Routes embedding vector requests.": "Routes embedding vector requests.",
+    "Routes image generation and editing requests.": "Routes image generation and editing requests.",
+    "Routes speech, transcription, translation, or music requests.": "Routes speech, transcription, translation, or music requests.",
+    "Routes search reranking requests.": "Routes search reranking requests.",
+    "Routes asynchronous video generation tasks.": "Routes asynchronous video generation tasks.",
+    "Fetches model lists directly from the upstream provider.": "Fetches model lists directly from the upstream provider.",
+    "chat/completions": "Chat / completions",
+    "responses": "Responses API",
+    "Gemini native": "Gemini native API",
+    "embeddings": "Embeddings",
+    "images": "Images",
+    "audio": "Audio",
+    "rerank": "Rerank",
+    "video tasks": "Video tasks",
+    "model discovery": "Model discovery",
+    "Native support": "Native support",
+    "Compatible support": "Compatible support",
+    "Configurable": "Configurable",
+    "Limited support": "Limited support",
+    "Not supported": "Not supported",
+    "This JSON configuration must be a JSON object.": "This JSON configuration must be a JSON object.",
+    "Vertex AI region configuration must be a JSON object containing a default field.": "Vertex AI region configuration must be a JSON object containing a default field.",
+    "Vertex AI region configuration must contain a default field.": "Vertex AI region configuration must contain a default field.",
+    "Vertex AI API Key mode does not support batch creation.": "Vertex AI API Key mode does not support batch creation.",
+    "Vertex AI service account credentials must be valid JSON.": "Vertex AI service account credentials must be valid JSON.",
+    "Codex channels do not support batch creation.": "Codex channels do not support batch creation.",
+    "Codex credentials must contain access_token and account_id.": "Codex credentials must contain access_token and account_id.",
+    "Codex credentials must be a valid JSON object.": "Codex credentials must be a valid JSON object.",
+    "Video task protocol settings are only available for channel types that support video tasks.": "Video task protocol settings are only available for channel types that support video tasks.",
+    "Video task submission path is required.": "Video task submission path is required.",
+    "Video task query path is required.": "Video task query path is required.",
+    "Video task query path must contain {task_id}, {operation_name}, or {upstream_task_id}.": "Video task query path must contain {task_id}, {operation_name}, or {upstream_task_id}.",
+    "Task ID response path is required when the full task protocol is enabled.": "Task ID response path is required when the full task protocol is enabled.",
+    "Status response path is required when the full task protocol is enabled.": "Status response path is required when the full task protocol is enabled.",
+    "Without a result URL path, completed video tasks may not return a content URL.": "Without a result URL path, completed video tasks may not return a content URL.",
+    "API Key is required when creating a channel.": "API Key is required when creating a channel.",
+    "At least one model must be published for this channel.": "At least one model must be published for this channel.",
+    "Base URL is required for this channel type.": "Base URL is required for this channel type.",
+    "Base URL should not usually end with /v1 because MAX API appends the upstream path automatically.": "Base URL should not usually end with /v1 because MAX API appends the upstream path automatically.",
+    "Additional upstream configuration is required for this channel type.": "Additional upstream configuration is required for this channel type.",
+    "Upstream model discovery is not supported for this channel type and will not run.": "Upstream model discovery is not supported for this channel type and will not run.",
     "_copy": "_copy",
     "，": ", ",
     ", and": ", and",
