+++ web/default/src/i18n/locales/fr.json
@@ -3,6 +3,58 @@
     "360": "360",
     "1000": "1000",
     "10000": "10000",
+    "{{supported}}/{{total}} available": "{{supported}}/{{total}} disponibles",
+    "Shows the request APIs, task capabilities, and operational features supported by this channel type.": "Affiche les API de requête, les fonctionnalités de tâches et les fonctions opérationnelles prises en charge par ce type de canal.",
+    "{{count}} blocking issue(s)": "{{count}} problème(s) bloquant(s)",
+    "Configuration valid": "Configuration valide",
+    "{{count}} warning(s)": "{{count}} avertissement(s)",
+    "Configuration validation": "Validation de la configuration",
+    "No blocking configuration issues were detected for this channel type.": "Aucun problème de configuration bloquant n'a été détecté pour ce type de canal.",
+    "Routes OpenAI-compatible chat and completion requests.": "Achemine les requêtes de chat et de complétion compatibles OpenAI.",
+    "Routes OpenAI Responses API requests.": "Achemine les requêtes de l'API OpenAI Responses.",
+    "Routes Anthropic Messages-compatible requests.": "Achemine les requêtes compatibles avec Anthropic Messages.",
+    "Routes Gemini native /v1beta model requests.": "Achemine les requêtes de modèles Gemini natifs /v1beta.",
+    "Routes embedding vector requests.": "Achemine les requêtes de vecteurs d'embedding.",
+    "Routes image generation and editing requests.": "Achemine les requêtes de génération et de modification d'images.",
+    "Routes speech, transcription, translation, or music requests.": "Achemine les requêtes de parole, transcription, traduction ou musique.",
+    "Routes search reranking requests.": "Achemine les requêtes de reclassement de recherche.",
+    "Routes asynchronous video generation tasks.": "Achemine les tâches asynchrones de génération vidéo.",
+    "Fetches model lists directly from the upstream provider.": "Récupère les listes de modèles directement auprès du fournisseur en amont.",
+    "chat/completions": "Chat / complétions",
+    "responses": "API Responses",
+    "Gemini native": "API native Gemini",
+    "embeddings": "Vectorisation",
+    "images": "Images",
+    "audio": "Audio",
+    "rerank": "Réordonnancement",
+    "video tasks": "Tâches vidéo",
+    "model discovery": "Découverte des modèles",
+    "Native support": "Prise en charge native",
+    "Compatible support": "Prise en charge compatible",
+    "Configurable": "Configurable",
+    "Limited support": "Prise en charge limitée",
+    "Not supported": "Non pris en charge",
+    "This JSON configuration must be a JSON object.": "Cette configuration JSON doit être un objet JSON.",
+    "Vertex AI region configuration must be a JSON object containing a default field.": "La configuration de région Vertex AI doit être un objet JSON contenant un champ default.",
+    "Vertex AI region configuration must contain a default field.": "La configuration de région Vertex AI doit contenir un champ default.",
+    "Vertex AI API Key mode does not support batch creation.": "Le mode API Key de Vertex AI ne prend pas en charge la création par lots.",
+    "Vertex AI service account credentials must be valid JSON.": "Les identifiants de compte de service Vertex AI doivent être un JSON valide.",
+    "Codex channels do not support batch creation.": "Les canaux Codex ne prennent pas en charge la création par lots.",
+    "Codex credentials must contain access_token and account_id.": "Les identifiants Codex doivent contenir access_token et account_id.",
+    "Codex credentials must be a valid JSON object.": "Les identifiants Codex doivent être un objet JSON valide.",
+    "Video task protocol settings are only available for channel types that support video tasks.": "Les paramètres du protocole de tâche vidéo sont réservés aux types de canaux prenant en charge les tâches vidéo.",
+    "Video task submission path is required.": "Le chemin de soumission des tâches vidéo est obligatoire.",
+    "Video task query path is required.": "Le chemin de requête des tâches vidéo est obligatoire.",
+    "Video task query path must contain {task_id}, {operation_name}, or {upstream_task_id}.": "Le chemin de requête des tâches vidéo doit contenir {task_id}, {operation_name} ou {upstream_task_id}.",
+    "Task ID response path is required when the full task protocol is enabled.": "Le chemin de réponse de l'ID de tâche est obligatoire lorsque le protocole complet est activé.",
+    "Status response path is required when the full task protocol is enabled.": "Le chemin de réponse du statut est obligatoire lorsque le protocole complet est activé.",
+    "Without a result URL path, completed video tasks may not return a content URL.": "Sans chemin d'URL de résultat, les tâches vidéo terminées peuvent ne pas renvoyer d'URL de contenu.",
+    "API Key is required when creating a channel.": "La clé API est obligatoire lors de la création d'un canal.",
+    "At least one model must be published for this channel.": "Au moins un modèle doit être publié pour ce canal.",
+    "Base URL is required for this channel type.": "L'URL de base est obligatoire pour ce type de canal.",
+    "Base URL should not usually end with /v1 because MAX API appends the upstream path automatically.": "L'URL de base ne doit généralement pas se terminer par /v1, car MAX API ajoute automatiquement le chemin en amont.",
+    "Additional upstream configuration is required for this channel type.": "Une configuration en amont supplémentaire est obligatoire pour ce type de canal.",
+    "Upstream model discovery is not supported for this channel type and will not run.": "La découverte des modèles en amont n'est pas prise en charge pour ce type de canal et ne sera pas exécutée.",
     "_copy": "_copie",
     "，": ", ",
     ", and": ", et",
