+++ web/default/src/i18n/locales/ja.json
@@ -3,6 +3,58 @@
     "360": "360",
     "1000": "1000",
     "10000": "10000",
+    "{{supported}}/{{total}} available": "{{supported}}/{{total}} 件利用可能",
+    "Shows the request APIs, task capabilities, and operational features supported by this channel type.": "このチャネル種別が対応するリクエスト API、タスク機能、運用機能を表示します。",
+    "{{count}} blocking issue(s)": "ブロック中の問題 {{count}} 件",
+    "Configuration valid": "設定は有効です",
+    "{{count}} warning(s)": "警告 {{count}} 件",
+    "Configuration validation": "設定の検証",
+    "No blocking configuration issues were detected for this channel type.": "このチャネル種別では、設定を妨げる問題は検出されませんでした。",
+    "Routes OpenAI-compatible chat and completion requests.": "OpenAI 互換のチャットおよび補完リクエストをルーティングします。",
+    "Routes OpenAI Responses API requests.": "OpenAI Responses API リクエストをルーティングします。",
+    "Routes Anthropic Messages-compatible requests.": "Anthropic Messages 互換リクエストをルーティングします。",
+    "Routes Gemini native /v1beta model requests.": "Gemini ネイティブ /v1beta モデルリクエストをルーティングします。",
+    "Routes embedding vector requests.": "埋め込みベクトルリクエストをルーティングします。",
+    "Routes image generation and editing requests.": "画像生成および編集リクエストをルーティングします。",
+    "Routes speech, transcription, translation, or music requests.": "音声、文字起こし、翻訳、音楽リクエストをルーティングします。",
+    "Routes search reranking requests.": "検索リランキングリクエストをルーティングします。",
+    "Routes asynchronous video generation tasks.": "非同期動画生成タスクをルーティングします。",
+    "Fetches model lists directly from the upstream provider.": "アップストリームプロバイダーからモデル一覧を直接取得します。",
+    "chat/completions": "チャット / 補完 API",
+    "responses": "レスポンス API",
+    "Gemini native": "Gemini ネイティブ API",
+    "embeddings": "埋め込み",
+    "images": "画像",
+    "audio": "音声",
+    "rerank": "再ランキング",
+    "video tasks": "動画タスク",
+    "model discovery": "モデル検出",
+    "Native support": "ネイティブ対応",
+    "Compatible support": "互換対応",
+    "Configurable": "設定可能",
+    "Limited support": "限定対応",
+    "Not supported": "非対応",
+    "This JSON configuration must be a JSON object.": "この JSON 設定は JSON オブジェクトである必要があります。",
+    "Vertex AI region configuration must be a JSON object containing a default field.": "Vertex AI のリージョン設定は default フィールドを含む JSON オブジェクトである必要があります。",
+    "Vertex AI region configuration must contain a default field.": "Vertex AI のリージョン設定には default フィールドが必要です。",
+    "Vertex AI API Key mode does not support batch creation.": "Vertex AI の API Key モードは一括作成に対応していません。",
+    "Vertex AI service account credentials must be valid JSON.": "Vertex AI のサービスアカウント認証情報は有効な JSON である必要があります。",
+    "Codex channels do not support batch creation.": "Codex チャネルは一括作成に対応していません。",
+    "Codex credentials must contain access_token and account_id.": "Codex の認証情報には access_token と account_id が必要です。",
+    "Codex credentials must be a valid JSON object.": "Codex の認証情報は有効な JSON オブジェクトである必要があります。",
+    "Video task protocol settings are only available for channel types that support video tasks.": "動画タスクプロトコル設定は、動画タスクに対応するチャネル種別でのみ使用できます。",
+    "Video task submission path is required.": "動画タスクの送信パスは必須です。",
+    "Video task query path is required.": "動画タスクの照会パスは必須です。",
+    "Video task query path must contain {task_id}, {operation_name}, or {upstream_task_id}.": "動画タスクの照会パスには {task_id}、{operation_name}、{upstream_task_id} のいずれかが必要です。",
+    "Task ID response path is required when the full task protocol is enabled.": "完全なタスクプロトコルを有効にする場合、タスク ID のレスポンスパスは必須です。",
+    "Status response path is required when the full task protocol is enabled.": "完全なタスクプロトコルを有効にする場合、ステータスのレスポンスパスは必須です。",
+    "Without a result URL path, completed video tasks may not return a content URL.": "結果 URL のパスがない場合、完了した動画タスクからコンテンツ URL が返されない可能性があります。",
+    "API Key is required when creating a channel.": "チャネルの作成時には API Key が必要です。",
+    "At least one model must be published for this channel.": "このチャネルでは少なくとも 1 つのモデルを公開する必要があります。",
+    "Base URL is required for this channel type.": "このチャネル種別では Base URL が必要です。",
+    "Base URL should not usually end with /v1 because MAX API appends the upstream path automatically.": "MAX API がアップストリームパスを自動追加するため、Base URL は通常 /v1 で終わらせないでください。",
+    "Additional upstream configuration is required for this channel type.": "このチャネル種別では追加のアップストリーム設定が必要です。",
+    "Upstream model discovery is not supported for this channel type and will not run.": "このチャネル種別はアップストリームモデル検出に対応していないため、実行されません。",
     "_copy": "_copy",
     "，": ", ",
     ", and": "、および",
