+++ web/default/src/i18n/locales/ru.json
@@ -3,6 +3,58 @@
     "360": "360",
     "1000": "1000",
     "10000": "10000",
+    "{{supported}}/{{total}} available": "Доступно: {{supported}}/{{total}}",
+    "Shows the request APIs, task capabilities, and operational features supported by this channel type.": "Показывает API запросов, возможности задач и рабочие функции, поддерживаемые этим типом канала.",
+    "{{count}} blocking issue(s)": "Блокирующих проблем: {{count}}",
+    "Configuration valid": "Конфигурация корректна",
+    "{{count}} warning(s)": "Предупреждений: {{count}}",
+    "Configuration validation": "Проверка конфигурации",
+    "No blocking configuration issues were detected for this channel type.": "Для этого типа канала не обнаружено блокирующих проблем конфигурации.",
+    "Routes OpenAI-compatible chat and completion requests.": "Маршрутизирует OpenAI-совместимые запросы чата и дополнения.",
+    "Routes OpenAI Responses API requests.": "Маршрутизирует запросы OpenAI Responses API.",
+    "Routes Anthropic Messages-compatible requests.": "Маршрутизирует запросы, совместимые с Anthropic Messages.",
+    "Routes Gemini native /v1beta model requests.": "Маршрутизирует запросы к моделям через нативный Gemini /v1beta.",
+    "Routes embedding vector requests.": "Маршрутизирует запросы векторных представлений.",
+    "Routes image generation and editing requests.": "Маршрутизирует запросы генерации и редактирования изображений.",
+    "Routes speech, transcription, translation, or music requests.": "Маршрутизирует запросы речи, транскрибации, перевода или музыки.",
+    "Routes search reranking requests.": "Маршрутизирует запросы переранжирования поиска.",
+    "Routes asynchronous video generation tasks.": "Маршрутизирует асинхронные задачи генерации видео.",
+    "Fetches model lists directly from the upstream provider.": "Получает список моделей непосредственно от вышестоящего провайдера.",
+    "chat/completions": "Чат / дополнения API",
+    "responses": "API ответов",
+    "Gemini native": "Нативный API Gemini",
+    "embeddings": "Векторные представления",
+    "images": "Изображения",
+    "audio": "Аудио",
+    "rerank": "Реранжирование",
+    "video tasks": "Видеозадачи",
+    "model discovery": "Обнаружение моделей",
+    "Native support": "Нативная поддержка",
+    "Compatible support": "Совместимая поддержка",
+    "Configurable": "Настраивается",
+    "Limited support": "Ограниченная поддержка",
+    "Not supported": "Не поддерживается",
+    "This JSON configuration must be a JSON object.": "Эта конфигурация JSON должна быть объектом JSON.",
+    "Vertex AI region configuration must be a JSON object containing a default field.": "Конфигурация региона Vertex AI должна быть объектом JSON с полем default.",
+    "Vertex AI region configuration must contain a default field.": "Конфигурация региона Vertex AI должна содержать поле default.",
+    "Vertex AI API Key mode does not support batch creation.": "Режим API Key в Vertex AI не поддерживает пакетное создание.",
+    "Vertex AI service account credentials must be valid JSON.": "Учетные данные сервисного аккаунта Vertex AI должны быть корректным JSON.",
+    "Codex channels do not support batch creation.": "Каналы Codex не поддерживают пакетное создание.",
+    "Codex credentials must contain access_token and account_id.": "Учетные данные Codex должны содержать access_token и account_id.",
+    "Codex credentials must be a valid JSON object.": "Учетные данные Codex должны быть корректным объектом JSON.",
+    "Video task protocol settings are only available for channel types that support video tasks.": "Настройки протокола видеозадач доступны только для типов каналов, поддерживающих видеозадачи.",
+    "Video task submission path is required.": "Путь отправки видеозадачи обязателен.",
+    "Video task query path is required.": "Путь запроса видеозадачи обязателен.",
+    "Video task query path must contain {task_id}, {operation_name}, or {upstream_task_id}.": "Путь запроса видеозадачи должен содержать {task_id}, {operation_name} или {upstream_task_id}.",
+    "Task ID response path is required when the full task protocol is enabled.": "При включенном полном протоколе задач требуется путь к ID задачи в ответе.",
+    "Status response path is required when the full task protocol is enabled.": "При включенном полном протоколе задач требуется путь к статусу в ответе.",
+    "Without a result URL path, completed video tasks may not return a content URL.": "Без пути к URL результата завершенные видеозадачи могут не вернуть URL содержимого.",
+    "API Key is required when creating a channel.": "При создании канала требуется API Key.",
+    "At least one model must be published for this channel.": "Для этого канала необходимо опубликовать хотя бы одну модель.",
+    "Base URL is required for this channel type.": "Для этого типа канала требуется Base URL.",
+    "Base URL should not usually end with /v1 because MAX API appends the upstream path automatically.": "Base URL обычно не должен заканчиваться на /v1, поскольку MAX API автоматически добавляет вышестоящий путь.",
+    "Additional upstream configuration is required for this channel type.": "Для этого типа канала требуется дополнительная конфигурация вышестоящего сервиса.",
+    "Upstream model discovery is not supported for this channel type and will not run.": "Обнаружение моделей вышестоящего сервиса не поддерживается для этого типа канала и выполняться не будет.",
     "_copy": "_копировать",
     "，": ", ",
     ", and": ", и",
