+++ web/default/src/i18n/locales/vi.json
@@ -3,6 +3,58 @@
     "360": "360",
     "1000": "1000",
     "10000": "10000",
+    "{{supported}}/{{total}} available": "Khả dụng {{supported}}/{{total}}",
+    "Shows the request APIs, task capabilities, and operational features supported by this channel type.": "Hiển thị API yêu cầu, khả năng tác vụ và tính năng vận hành được loại kênh này hỗ trợ.",
+    "{{count}} blocking issue(s)": "{{count}} vấn đề chặn",
+    "Configuration valid": "Cấu hình hợp lệ",
+    "{{count}} warning(s)": "{{count}} cảnh báo",
+    "Configuration validation": "Xác thực cấu hình",
+    "No blocking configuration issues were detected for this channel type.": "Không phát hiện vấn đề cấu hình chặn nào cho loại kênh này.",
+    "Routes OpenAI-compatible chat and completion requests.": "Định tuyến yêu cầu trò chuyện và hoàn thành tương thích OpenAI.",
+    "Routes OpenAI Responses API requests.": "Định tuyến yêu cầu OpenAI Responses API.",
+    "Routes Anthropic Messages-compatible requests.": "Định tuyến yêu cầu tương thích Anthropic Messages.",
+    "Routes Gemini native /v1beta model requests.": "Định tuyến yêu cầu mô hình Gemini gốc /v1beta.",
+    "Routes embedding vector requests.": "Định tuyến yêu cầu vector embedding.",
+    "Routes image generation and editing requests.": "Định tuyến yêu cầu tạo và chỉnh sửa ảnh.",
+    "Routes speech, transcription, translation, or music requests.": "Định tuyến yêu cầu giọng nói, phiên âm, dịch thuật hoặc âm nhạc.",
+    "Routes search reranking requests.": "Định tuyến yêu cầu xếp hạng lại kết quả tìm kiếm.",
+    "Routes asynchronous video generation tasks.": "Định tuyến tác vụ tạo video bất đồng bộ.",
+    "Fetches model lists directly from the upstream provider.": "Lấy danh sách mô hình trực tiếp từ nhà cung cấp thượng nguồn.",
+    "chat/completions": "Trò chuyện / hoàn thành API",
+    "responses": "API phản hồi",
+    "Gemini native": "API Gemini gốc",
+    "embeddings": "Nhúng",
+    "images": "Hình ảnh",
+    "audio": "Âm thanh",
+    "rerank": "Xếp hạng lại",
+    "video tasks": "Tác vụ video",
+    "model discovery": "Khám phá mô hình",
+    "Native support": "Hỗ trợ gốc",
+    "Compatible support": "Hỗ trợ tương thích",
+    "Configurable": "Có thể cấu hình",
+    "Limited support": "Hỗ trợ hạn chế",
+    "Not supported": "Không hỗ trợ",
+    "This JSON configuration must be a JSON object.": "Cấu hình JSON này phải là một đối tượng JSON.",
+    "Vertex AI region configuration must be a JSON object containing a default field.": "Cấu hình vùng Vertex AI phải là đối tượng JSON chứa trường default.",
+    "Vertex AI region configuration must contain a default field.": "Cấu hình vùng Vertex AI phải chứa trường default.",
+    "Vertex AI API Key mode does not support batch creation.": "Chế độ API Key của Vertex AI không hỗ trợ tạo hàng loạt.",
+    "Vertex AI service account credentials must be valid JSON.": "Thông tin xác thực tài khoản dịch vụ Vertex AI phải là JSON hợp lệ.",
+    "Codex channels do not support batch creation.": "Kênh Codex không hỗ trợ tạo hàng loạt.",
+    "Codex credentials must contain access_token and account_id.": "Thông tin xác thực Codex phải chứa access_token và account_id.",
+    "Codex credentials must be a valid JSON object.": "Thông tin xác thực Codex phải là một đối tượng JSON hợp lệ.",
+    "Video task protocol settings are only available for channel types that support video tasks.": "Cài đặt giao thức tác vụ video chỉ khả dụng cho loại kênh hỗ trợ tác vụ video.",
+    "Video task submission path is required.": "Đường dẫn gửi tác vụ video là bắt buộc.",
+    "Video task query path is required.": "Đường dẫn truy vấn tác vụ video là bắt buộc.",
+    "Video task query path must contain {task_id}, {operation_name}, or {upstream_task_id}.": "Đường dẫn truy vấn tác vụ video phải chứa {task_id}, {operation_name} hoặc {upstream_task_id}.",
+    "Task ID response path is required when the full task protocol is enabled.": "Đường dẫn ID tác vụ trong phản hồi là bắt buộc khi bật giao thức tác vụ đầy đủ.",
+    "Status response path is required when the full task protocol is enabled.": "Đường dẫn trạng thái trong phản hồi là bắt buộc khi bật giao thức tác vụ đầy đủ.",
+    "Without a result URL path, completed video tasks may not return a content URL.": "Nếu không có đường dẫn URL kết quả, tác vụ video đã hoàn thành có thể không trả về URL nội dung.",
+    "API Key is required when creating a channel.": "API Key là bắt buộc khi tạo kênh.",
+    "At least one model must be published for this channel.": "Phải phát hành ít nhất một mô hình cho kênh này.",
+    "Base URL is required for this channel type.": "Base URL là bắt buộc đối với loại kênh này.",
+    "Base URL should not usually end with /v1 because MAX API appends the upstream path automatically.": "Base URL thường không nên kết thúc bằng /v1 vì MAX API tự động nối đường dẫn thượng nguồn.",
+    "Additional upstream configuration is required for this channel type.": "Loại kênh này yêu cầu cấu hình thượng nguồn bổ sung.",
+    "Upstream model discovery is not supported for this channel type and will not run.": "Loại kênh này không hỗ trợ khám phá mô hình thượng nguồn nên tính năng sẽ không chạy.",
     "_copy": "_bản sao",
     "，": ", ",
     ", and": ", và",
