+++ web/default/src/i18n/locales/zh.json
@@ -3,6 +3,58 @@
     "360": "360",
     "1000": "1000",
     "10000": "10000",
+    "{{supported}}/{{total}} available": "{{supported}}/{{total}} 项可用",
+    "Shows the request APIs, task capabilities, and operational features supported by this channel type.": "展示当前渠道类型支持的请求接口、任务能力和运营功能。",
+    "{{count}} blocking issue(s)": "{{count}} 个阻断问题",
+    "Configuration valid": "配置一致",
+    "{{count}} warning(s)": "{{count}} 个警告",
+    "Configuration validation": "配置校验",
+    "No blocking configuration issues were detected for this channel type.": "当前渠道类型未检测到阻断性配置问题。",
+    "Routes OpenAI-compatible chat and completion requests.": "用于 OpenAI 兼容的聊天与补全接口请求。",
+    "Routes OpenAI Responses API requests.": "用于 OpenAI Responses 风格接口请求。",
+    "Routes Anthropic Messages-compatible requests.": "用于 Anthropic Messages 兼容接口请求。",
+    "Routes Gemini native /v1beta model requests.": "用于 Gemini 原生 /v1beta 模型路由。",
+    "Routes embedding vector requests.": "用于 embedding 向量接口请求。",
+    "Routes image generation and editing requests.": "用于图像生成或图像编辑接口请求。",
+    "Routes speech, transcription, translation, or music requests.": "用于语音、转写、翻译或音乐任务接口。",
+    "Routes search reranking requests.": "用于搜索重排序接口请求。",
+    "Routes asynchronous video generation tasks.": "用于异步视频生成任务接口。",
+    "Fetches model lists directly from the upstream provider.": "用于直接从上游拉取模型列表。",
+    "chat/completions": "聊天 / 补全 API",
+    "responses": "响应 API",
+    "Gemini native": "Gemini 原生 API",
+    "embeddings": "向量嵌入",
+    "images": "图像",
+    "audio": "音频",
+    "rerank": "重排序",
+    "video tasks": "视频任务",
+    "model discovery": "模型发现",
+    "Native support": "原生支持",
+    "Compatible support": "兼容支持",
+    "Configurable": "可配置",
+    "Limited support": "有限支持",
+    "Not supported": "不支持",
+    "This JSON configuration must be a JSON object.": "此 JSON 配置必须是 JSON 对象。",
+    "Vertex AI region configuration must be a JSON object containing a default field.": "Vertex AI 区域配置必须是包含 default 字段的 JSON 对象。",
+    "Vertex AI region configuration must contain a default field.": "Vertex AI 区域配置必须包含 default 字段。",
+    "Vertex AI API Key mode does not support batch creation.": "Vertex AI 的 API Key 模式不支持批量创建。",
+    "Vertex AI service account credentials must be valid JSON.": "Vertex AI 服务账号密钥必须是有效的 JSON。",
+    "Codex channels do not support batch creation.": "Codex 渠道不支持批量创建。",
+    "Codex credentials must contain access_token and account_id.": "Codex 凭证必须包含 access_token 和 account_id 字段。",
+    "Codex credentials must be a valid JSON object.": "Codex 凭证必须是有效的 JSON 对象。",
+    "Video task protocol settings are only available for channel types that support video tasks.": "视频任务协议设置仅适用于支持视频任务的渠道类型。",
+    "Video task submission path is required.": "视频任务提交路径不能为空。",
+    "Video task query path is required.": "视频任务查询路径不能为空。",
+    "Video task query path must contain {task_id}, {operation_name}, or {upstream_task_id}.": "视频任务查询路径必须包含 {task_id}、{operation_name} 或 {upstream_task_id}。",
+    "Task ID response path is required when the full task protocol is enabled.": "启用完整任务协议时，必须填写任务 ID 响应路径。",
+    "Status response path is required when the full task protocol is enabled.": "启用完整任务协议时，必须填写状态响应路径。",
+    "Without a result URL path, completed video tasks may not return a content URL.": "未配置结果 URL 路径，视频任务完成后可能无法返回内容地址。",
+    "API Key is required when creating a channel.": "新建渠道时 API Key 不能为空。",
+    "At least one model must be published for this channel.": "该渠道至少需要发布一个模型。",
+    "Base URL is required for this channel type.": "当前渠道类型需要填写 Base URL。",
+    "Base URL should not usually end with /v1 because MAX API appends the upstream path automatically.": "Base URL 通常不应以 /v1 结尾，MAX API 会自动拼接上游路径。",
+    "Additional upstream configuration is required for this channel type.": "当前渠道类型需要填写额外的上游配置。",
+    "Upstream model discovery is not supported for this channel type and will not run.": "当前渠道类型不支持上游模型发现，该功能不会运行。",
     "_copy": "_复制",
     "，": "，",
     ", and": "，和",
